refactor: extract template in file then dialog logic in a service
