+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-details-page.html
@@ -1,0 +1,147 @@
+<div class="flex flex-col gap-8 min-w-0" data-testid="budget-detail-page">
+  @if (store.isLoading()) {
+  <pulpe-base-loading
+    message="Préparation de ton budget..."
+    size="large"
+    [fullHeight]="true"
+    testId="budget-details-loading"
+  ></pulpe-base-loading>
+  } @else if (store.error()) {
+  <mat-card class="bg-error-container/50 border-0" appearance="outlined">
+    <mat-card-content>
+      <div class="flex items-center gap-3 text-on-error-container py-2">
+        <mat-icon>error_outline</mat-icon>
+        <span class="text-body-large"
+          >Impossible de charger ton budget — réessaie</span
+        >
+      </div>
+    </mat-card-content>
+  </mat-card>
+  } @else if (store.budgetDetails()) { @let budget = store.budgetDetails()!;
+  @let budgetLines = store.displayBudgetLines(); @let transactions =
+  budget.transactions;
+
+  <!-- Header: Expressive with breathing room -->
+  <header class="relative">
+    <div class="flex items-start gap-3 sm:gap-4 min-w-0">
+      <button
+        matIconButton
+        (click)="navigateBack()"
+        aria-label="Retour aux budgets"
+        data-testid="back-button"
+        class="flex-shrink-0 -ml-2"
+      >
+        <mat-icon>arrow_back</mat-icon>
+      </button>
+      <div class="flex-1 min-w-0 pt-1">
+        <h1
+          class="text-display-small sm:text-display-medium mb-2 truncate capitalize"
+        >
+          {{ displayName() }}
+        </h1>
+        @if (periodDisplay()) {
+        <p
+          class="text-body-large text-on-surface-variant"
+          data-testid="budget-period-display"
+        >
+          {{ periodDisplay() }}
+        </p>
+        } @if (budget.description) {
+        <p class="text-body-medium text-on-surface-variant mt-1">
+          {{ budget.description }}
+        </p>
+        }
+      </div>
+      <button
+        matIconButton
+        (click)="startPageTour()"
+        matTooltip="Découvrir cette page"
+        aria-label="Aide"
+        data-testid="help-button"
+        class="flex-shrink-0 -mr-2"
+      >
+        <mat-icon>help_outline</mat-icon>
+      </button>
+    </div>
+  </header>
+
+  <!-- Financial Overview -->
+  <pulpe-budget-financial-overview
+    [budgetLines]="budgetLines"
+    [transactions]="transactions"
+    [realizedBalance]="store.realizedBalance()"
+    [realizedExpenses]="store.realizedExpenses()"
+    [checkedCount]="store.checkedItemsCount()"
+    [totalCount]="store.totalItemsCount()"
+    data-tour="financial-overview"
+  />
+
+  <!-- Budget Items -->
+  <pulpe-budget-items
+    [budgetLines]="budgetLines"
+    [transactions]="transactions"
+    (update)="handleUpdateBudgetLine($event)"
+    (delete)="handleDeleteItem($event)"
+    (deleteTransaction)="handleDeleteItem($event)"
+    (add)="openAddBudgetLineDialog()"
+    (viewAllocatedTransactions)="openAllocatedTransactionsDialog($event)"
+    (createAllocatedTransaction)="
+        openCreateAllocatedTransactionDialog($event)
+      "
+    (resetFromTemplate)="handleResetFromTemplate($event)"
+    (toggleCheck)="handleToggleCheck($event)"
+    (toggleTransactionCheck)="handleToggleTransactionCheck($event)"
+    data-tour="budget-table"
+  />
+
+  @if (isDevMode) {
+  <!-- Budget Info Card -->
+  <mat-card appearance="outlined">
+    <mat-card-header>
+      <div mat-card-avatar>
+        <div
+          class="flex justify-center items-center size-11 bg-primary-container rounded-full"
+        >
+          <mat-icon class="text-on-primary-container">calendar_month</mat-icon>
+        </div>
+      </div>
+      <mat-card-title>Informations du budget</mat-card-title>
+      <mat-card-subtitle>Détails et métadonnées</mat-card-subtitle>
+    </mat-card-header>
+    <mat-card-content>
+      <div class="grid grid-cols-1 md:grid-cols-2 gap-4 mt-4">
+        <div>
+          <div class="text-label-medium text-on-surface-variant">Période</div>
+          <p class="text-body-large">{{ displayName() }}</p>
+        </div>
+        <div>
+          <div class="text-label-medium text-on-surface-variant">Créé le</div>
+          <p class="text-body-large">
+            {{ budget.createdAt | date: 'short' : '' : 'fr-CH' }}
+          </p>
+        </div>
+        <div>
+          <div class="text-label-medium text-on-surface-variant">
+            Dernière modification
+          </div>
+          <p class="text-body-large">
+            {{ budget.updatedAt | date: 'short' : '' : 'fr-CH' }}
+          </p>
+        </div>
+        <div>
+          <div class="text-label-medium text-on-surface-variant">
+            ID du budget
+          </div>
+          <p class="text-body-small font-mono text-on-surface-variant">
+            {{ budget.id }}
+          </p>
+        </div>
+      </div>
+    </mat-card-content>
+  </mat-card>
+  } } @else {
+  <div class="flex justify-center items-center h-full">
+    <p class="text-body-large">Budget introuvable</p>
+  </div>
+  }
+</div>
