+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-details-page.ts
@@ -1,4 +1,3 @@
-import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import {
   afterNextRender,
   ChangeDetectionStrategy,
@@ -9,57 +8,35 @@ import {
   effect,
   isDevMode,
 } from '@angular/core';
-import { toSignal } from '@angular/core/rxjs-interop';
-import { firstValueFrom, map } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatBottomSheet } from '@angular/material/bottom-sheet';
-import { BaseLoading } from '@ui/loading';
-import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 import { DatePipe } from '@angular/common';
+import type { BudgetLineConsumption } from '@core/budget';
 import { Logger } from '@core/logging/logger';
 import { BreadcrumbState } from '@core/routing';
+import {
+  ProductTourService,
+  TOUR_START_DELAY,
+} from '@core/product-tour/product-tour.service';
 import { formatDate } from 'date-fns';
 import { frCH } from 'date-fns/locale';
+import { BaseLoading } from '@ui/loading';
 import { BudgetDetailsStore } from './store/budget-details-store';
 import { BudgetLineApi } from './budget-line-api/budget-line-api';
 import { BudgetItemsContainer } from './budget-items-container';
 import { BudgetFinancialOverview } from './budget-financial-overview';
-import {
-  AddBudgetLineDialog,
-  type BudgetLineDialogData,
-} from './create-budget-line/add-budget-line-dialog';
-import {
-  ConfirmationDialog,
-  type ConfirmationDialogData,
-} from '@ui/dialogs/confirmation-dialog';
+import { BudgetDetailsDialogService } from './budget-details-dialog.service';
 import {
   type BudgetLineCreate,
   type BudgetLineUpdate,
   type BudgetLine,
   type Transaction,
-  type TransactionCreate,
   formatBudgetPeriod,
 } from 'pulpe-shared';
-import type { BudgetLineConsumption } from '@core/budget';
-import {
-  AllocatedTransactionsDialog,
-  type AllocatedTransactionsDialogData,
-  type AllocatedTransactionsDialogResult,
-} from './allocated-transactions-dialog/allocated-transactions-dialog';
-import { AllocatedTransactionsBottomSheet } from './allocated-transactions-dialog/allocated-transactions-bottom-sheet';
-import {
-  CreateAllocatedTransactionDialog,
-  type CreateAllocatedTransactionDialogData,
-} from './create-allocated-transaction-dialog/create-allocated-transaction-dialog';
-import {
-  ProductTourService,
-  TOUR_START_DELAY,
-} from '@core/product-tour/product-tour.service';
 import { UserSettingsApi } from '@core/user-settings/user-settings-api';
 
 @Component({
@@ -69,174 +46,14 @@ import { UserSettingsApi } from '@core/user-settings/user-settings-api';
     MatIconModule,
     MatButtonModule,
     MatSnackBarModule,
-    MatDialogModule,
     MatTooltipModule,
     DatePipe,
     BudgetItemsContainer,
     BudgetFinancialOverview,
     BaseLoading,
   ],
-  providers: [BudgetDetailsStore, BudgetLineApi],
-  template: `
-    <div class="flex flex-col gap-8 min-w-0" data-testid="budget-detail-page">
-      @if (store.isLoading()) {
-        <pulpe-base-loading
-          message="Préparation de ton budget..."
-          size="large"
-          [fullHeight]="true"
-          testId="budget-details-loading"
-        ></pulpe-base-loading>
-      } @else if (store.error()) {
-        <mat-card class="bg-error-container/50 border-0" appearance="outlined">
-          <mat-card-content>
-            <div class="flex items-center gap-3 text-on-error-container py-2">
-              <mat-icon>error_outline</mat-icon>
-              <span class="text-body-large"
-                >Impossible de charger ton budget — réessaie</span
-              >
-            </div>
-          </mat-card-content>
-        </mat-card>
-      } @else if (store.budgetDetails()) {
-        @let budget = store.budgetDetails()!;
-        @let budgetLines = store.displayBudgetLines();
-        @let transactions = budget.transactions;
-
-        <!-- Header: Expressive with breathing room -->
-        <header class="relative">
-          <div class="flex items-start gap-3 sm:gap-4 min-w-0">
-            <button
-              matIconButton
-              (click)="navigateBack()"
-              aria-label="Retour aux budgets"
-              data-testid="back-button"
-              class="flex-shrink-0 -ml-2"
-            >
-              <mat-icon>arrow_back</mat-icon>
-            </button>
-            <div class="flex-1 min-w-0 pt-1">
-              <h1
-                class="text-display-small sm:text-display-medium mb-2 truncate capitalize"
-              >
-                {{ displayName() }}
-              </h1>
-              @if (periodDisplay()) {
-                <p
-                  class="text-body-large text-on-surface-variant"
-                  data-testid="budget-period-display"
-                >
-                  {{ periodDisplay() }}
-                </p>
-              }
-              @if (budget.description) {
-                <p class="text-body-medium text-on-surface-variant mt-1">
-                  {{ budget.description }}
-                </p>
-              }
-            </div>
-            <button
-              matIconButton
-              (click)="startPageTour()"
-              matTooltip="Découvrir cette page"
-              aria-label="Aide"
-              data-testid="help-button"
-              class="flex-shrink-0 -mr-2"
-            >
-              <mat-icon>help_outline</mat-icon>
-            </button>
-          </div>
-        </header>
-
-        <!-- Financial Overview -->
-        <pulpe-budget-financial-overview
-          [budgetLines]="budgetLines"
-          [transactions]="transactions"
-          [realizedBalance]="store.realizedBalance()"
-          [realizedExpenses]="store.realizedExpenses()"
-          [checkedCount]="store.checkedItemsCount()"
-          [totalCount]="store.totalItemsCount()"
-          data-tour="financial-overview"
-        />
-
-        <!-- Budget Items -->
-        <pulpe-budget-items
-          [budgetLines]="budgetLines"
-          [transactions]="transactions"
-          (update)="handleUpdateBudgetLine($event)"
-          (delete)="handleDeleteItem($event)"
-          (deleteTransaction)="handleDeleteItem($event)"
-          (add)="openAddBudgetLineDialog()"
-          (viewAllocatedTransactions)="openAllocatedTransactionsDialog($event)"
-          (createAllocatedTransaction)="
-            openCreateAllocatedTransactionDialog($event)
-          "
-          (resetFromTemplate)="handleResetFromTemplate($event)"
-          (toggleCheck)="handleToggleCheck($event)"
-          (toggleTransactionCheck)="handleToggleTransactionCheck($event)"
-          data-tour="budget-table"
-        />
-
-        @if (isDevMode) {
-          <!-- Budget Info Card -->
-          <mat-card appearance="outlined">
-            <mat-card-header>
-              <div mat-card-avatar>
-                <div
-                  class="flex justify-center items-center size-11 bg-primary-container rounded-full"
-                >
-                  <mat-icon class="text-on-primary-container"
-                    >calendar_month</mat-icon
-                  >
-                </div>
-              </div>
-              <mat-card-title>Informations du budget</mat-card-title>
-              <mat-card-subtitle>Détails et métadonnées</mat-card-subtitle>
-            </mat-card-header>
-            <mat-card-content>
-              <div class="grid grid-cols-1 md:grid-cols-2 gap-4 mt-4">
-                <div>
-                  <div class="text-label-medium text-on-surface-variant">
-                    Période
-                  </div>
-                  <p class="text-body-large">
-                    {{ displayName() }}
-                  </p>
-                </div>
-                <div>
-                  <div class="text-label-medium text-on-surface-variant">
-                    Créé le
-                  </div>
-                  <p class="text-body-large">
-                    {{ budget.createdAt | date: 'short' : '' : 'fr-CH' }}
-                  </p>
-                </div>
-                <div>
-                  <div class="text-label-medium text-on-surface-variant">
-                    Dernière modification
-                  </div>
-                  <p class="text-body-large">
-                    {{ budget.updatedAt | date: 'short' : '' : 'fr-CH' }}
-                  </p>
-                </div>
-                <div>
-                  <div class="text-label-medium text-on-surface-variant">
-                    ID du budget
-                  </div>
-                  <p class="text-body-small font-mono text-on-surface-variant">
-                    {{ budget.id }}
-                  </p>
-                </div>
-              </div>
-            </mat-card-content>
-          </mat-card>
-        }
-      } @else {
-        <div class="flex justify-center items-center h-full">
-          <p class="text-body-large">Budget introuvable</p>
-        </div>
-      }
-    </div>
-  `,
+  providers: [BudgetDetailsStore, BudgetLineApi, BudgetDetailsDialogService],
+  templateUrl: './budget-details-page.html',
   styles: `
     :host {
       display: block;
@@ -247,23 +64,14 @@ import { UserSettingsApi } from '@core/user-settings/user-settings-api';
 export default class BudgetDetailsPage {
   readonly isDevMode = isDevMode();
   readonly store = inject(BudgetDetailsStore);
+  readonly #dialogService = inject(BudgetDetailsDialogService);
   readonly #router = inject(Router);
   readonly #route = inject(ActivatedRoute);
-  readonly #dialog = inject(MatDialog);
-  readonly #bottomSheet = inject(MatBottomSheet);
+  readonly #breadcrumbState = inject(BreadcrumbState);
+  readonly #productTourService = inject(ProductTourService);
   readonly #snackBar = inject(MatSnackBar);
   readonly #logger = inject(Logger);
-  readonly #productTourService = inject(ProductTourService);
-  readonly #breadcrumbState = inject(BreadcrumbState);
-  readonly #breakpointObserver = inject(BreakpointObserver);
   readonly #userSettingsApi = inject(UserSettingsApi);
-
-  readonly #isMobile = toSignal(
-    this.#breakpointObserver
-      .observe(Breakpoints.Handset)
-      .pipe(map((result) => result.matches)),
-    { initialValue: false },
-  );
 
   readonly id = input.required<string>();
 
@@ -327,15 +135,9 @@ export default class BudgetDetailsPage {
     const budget = this.store.budgetDetails();
     if (!budget) return;
 
-    const dialogRef = this.#dialog.open(AddBudgetLineDialog, {
-      data: {
-        budgetId: budget.id,
-      } satisfies BudgetLineDialogData,
-      width: '600px',
-      maxWidth: '90vw',
-    });
-
-    const budgetLine = await firstValueFrom(dialogRef.afterClosed());
+    const budgetLine = await this.#dialogService.openAddBudgetLineDialog(
+      budget.id,
+    );
     if (budgetLine) {
       this.handleCreateBudgetLine(budgetLine);
     }
@@ -358,7 +160,6 @@ export default class BudgetDetailsPage {
     const data = this.store.budgetDetails();
     if (!data) return;
 
-    // Find the item to determine if it's a budget line or transaction
     const budgetLine = data.budgetLines.find(
       (line: BudgetLine) => line.id === id,
     );
@@ -375,25 +176,14 @@ export default class BudgetDetailsPage {
     const title = isBudgetLine
       ? 'Supprimer cette prévision ?'
       : 'Supprimer cette transaction ?';
-    const message = isBudgetLine
-      ? 'Cette action est irréversible.'
-      : 'Cette action est irréversible.';
+    const message = 'Cette action est irréversible.';
 
-    const dialogRef = this.#dialog.open(ConfirmationDialog, {
-      data: {
-        title,
-        message,
-        confirmText: 'Supprimer',
-        confirmColor: 'warn',
-      } satisfies ConfirmationDialogData,
-      width: '400px',
+    const confirmed = await this.#dialogService.confirmDelete({
+      title,
+      message,
     });
 
-    const confirmed = await firstValueFrom(dialogRef.afterClosed());
-
-    if (!confirmed) {
-      return;
-    }
+    if (!confirmed) return;
 
     if (isBudgetLine) {
       await this.store.deleteBudgetLine(id);
@@ -407,65 +197,29 @@ export default class BudgetDetailsPage {
     }
   }
 
-  /**
-   * Open dialog or bottom sheet to view allocated transactions for a budget line
-   */
   async openAllocatedTransactionsDialog(event: {
     budgetLine: BudgetLine;
     consumption: BudgetLineConsumption;
   }): Promise<void> {
-    const data: AllocatedTransactionsDialogData = {
-      budgetLine: event.budgetLine,
-      consumption: event.consumption,
-    };
-
-    let result: AllocatedTransactionsDialogResult | undefined;
-
-    if (this.#isMobile()) {
-      // Mobile: use bottom sheet
-      const bottomSheetRef = this.#bottomSheet.open(
-        AllocatedTransactionsBottomSheet,
-        { data },
-      );
-      result = await firstValueFrom(bottomSheetRef.afterDismissed());
-    } else {
-      // Desktop: use dialog
-      const dialogRef = this.#dialog.open(AllocatedTransactionsDialog, {
-        data,
-        width: '800px',
-        maxWidth: '95vw',
-      });
-      result = await firstValueFrom(dialogRef.afterClosed());
-    }
+    const result =
+      await this.#dialogService.openAllocatedTransactionsDialog(event);
 
     if (!result) return;
 
     if (result.action === 'add') {
-      // Open create transaction dialog
       this.openCreateAllocatedTransactionDialog(event.budgetLine);
     } else if (result.action === 'delete' && result.transaction) {
-      // Delete transaction with confirmation
       await this.handleDeleteTransaction(result.transaction);
     }
   }
 
-  /**
-   * Open dialog to create an allocated transaction
-   */
   async openCreateAllocatedTransactionDialog(
     budgetLine: BudgetLine,
   ): Promise<void> {
-    const dialogRef = this.#dialog.open(CreateAllocatedTransactionDialog, {
-      data: {
+    const transaction =
+      await this.#dialogService.openCreateAllocatedTransactionDialog(
         budgetLine,
-      } satisfies CreateAllocatedTransactionDialogData,
-      width: '600px',
-      maxWidth: '90vw',
-    });
-
-    const transaction: TransactionCreate | undefined = await firstValueFrom(
-      dialogRef.afterClosed(),
-    );
+      );
 
     if (transaction) {
       await this.store.createAllocatedTransaction(transaction);
@@ -476,21 +230,11 @@ export default class BudgetDetailsPage {
     }
   }
 
-  /**
-   * Delete a transaction with confirmation dialog
-   */
   async handleDeleteTransaction(transaction: Transaction): Promise<void> {
-    const dialogRef = this.#dialog.open(ConfirmationDialog, {
-      data: {
-        title: 'Supprimer cette transaction ?',
-        message: `Tu vas supprimer « ${transaction.name} ». Cette action est irréversible.`,
-        confirmText: 'Supprimer',
-        confirmColor: 'warn',
-      } satisfies ConfirmationDialogData,
-      width: '400px',
+    const confirmed = await this.#dialogService.confirmDelete({
+      title: 'Supprimer cette transaction ?',
+      message: `Tu vas supprimer « ${transaction.name} ». Cette action est irréversible.`,
     });
-
-    const confirmed = await firstValueFrom(dialogRef.afterClosed());
 
     if (confirmed) {
       await this.store.deleteTransaction(transaction.id);
